backend-api/requirements.txt
algoliasearch-django>=3.0,<4.0
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
